eventmanager/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pytz==2024.1
PyYAML==6.0.2
setuptools==74.1.1
sqlparse==0.5.1
uritemplate==4.1.1
